refactor(darwin): remove C imports and update build constraints for non-Darwin platforms
