+++ README.md
@@ -4,7 +4,23 @@ This is a Go binding to the [ICU4X](https://github.com/unicode-org/icu4x) C FFI
 
 ## Features
 
-## Install
+- [x] Get Locale properties
+- [x] NumberFormatter
+- [ ] ListFormatter: join a list of items with ListInitType and ListLength
+- [ ] PluralRules
+- [ ] DateTimeFormatter
+
+## Installation
+
+### macOS
+
+```bash
+# or use latest commit, e.g:
+# go get github.com/Seeingu/icu4xgo@4ebc70a0d4977f56b26a2eb65d77267b2c57e8c4
+go get github.com/Seeingu/icu4xgo@v0.1.3
+```
+
+### Other OS
 
 Prerequisites
 
@@ -13,18 +29,10 @@ Prerequisites
 ```bash
 # or use latest commit, e.g:
 # go get github.com/Seeingu/icu4xgo@4ebc70a0d4977f56b26a2eb65d77267b2c57e8c4
-go get github.com/Seeingu/icu4xgo@v0.1.2
+go get github.com/Seeingu/icu4xgo@v0.1.3
 
 # If you encounter permission issues, run make with sudo
 cd `go list -f "{{.Dir}}" github.com/Seeingu/icu4xgo` && make rustlib
-```
-
-### Optional: Build from source
-
-2. Build and run tests
-
-```sh
-make all
 ```
 
 ## Usage
@@ -41,7 +49,7 @@ import (
 func main() {
 	locale := icu4xgo.NewLocale("zh-Hans-CN-u-ca-chinese-hc-h12")
 	fmt.Println(locale.BaseName()) // zh-Hans-CN
-	hc, _ := locale.HourCycle()
+	hc := locale.HourCycle()
 	fmt.Println(hc) // h12
 }
 ```
